requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.0.3
djangorestframework==3.16.0
idna==3.10
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0

